backend/requirements.txt
doppler-sdk
Faker
python-dotenv
fpdf2==2.7.8
Flask
flask-cors
sql_metadata
sqlalchemy==1.4.49
psycopg2-binary
rasa